Extract datetime picker from EditDateTimeCard.vue
